+++ src/commands/checks/index.ts
@@ -1,0 +1,20 @@
+export { runPrelaunchChecklist } from './prelaunchChecklist.js';
+export { runWalletCheck } from './walletCheck.js';
+export { runBalanceCheck } from './balanceCheck.js';
+export { runBuildCheck } from './buildCheck.js';
+export { runAntCheck } from './antCheck.js';
+export { runGitCheck } from './gitCheck.js';
+
+// Types for check results
+export interface CheckResult {
+  success: boolean;
+  message?: string;
+}
+
+export interface PrelaunchCheckResults {
+  wallet: CheckResult;
+  balance: CheckResult;
+  build: CheckResult;
+  ant: CheckResult;
+  git: CheckResult;
+}
