+++ src/commands/index.ts
@@ -1,0 +1,45 @@
+import { DeployArgs } from '../types.js';
+
+export async function handleCommands(argv: DeployArgs): Promise<void> {
+  if (argv['prelaunch-checklist']) {
+    // Will implement prelaunch checklist handler
+    return;
+  }
+
+  if (argv['check-wallet']) {
+    // Will implement wallet check handler
+    return;
+  }
+
+  if (argv['check-balances']) {
+    // Will implement balance check handler
+    return;
+  }
+
+  if (argv['check-build']) {
+    // Will implement build check handler
+    return;
+  }
+
+  if (argv['check-ant']) {
+    // Will implement ANT check handler
+    return;
+  }
+
+  if (argv['check-git']) {
+    // Will implement git check handler
+    return;
+  }
+
+  if (argv.launch) {
+    // Will implement launch handler
+    return;
+  }
+
+  if (argv['quick-launch']) {
+    // Will implement quick launch handler
+    return;
+  }
+
+  console.log('Please specify either --launch, --prelaunch-checklist, or --check-* flag');
+}
